backend/requirements.txt
mysql-connector-python==8.0.25
flask==1.1.2
flask-cors==3.0.10
flask-jwt-extended==4.3.0
numpy==1.18.5
pandas==1.0.5
redis==3.5.3
scipy==1.4.1
werkzeug==1.0.1
gunicorn==20.1.0
